feat(build): Add xlsx to external dependencies
Includes the `xlsx` library as an external dependency in the build process and in `index.html` for potential future use.

Removes the `sed` command from the build script as it was intended to replace the `.tsx` extension with `.js` in `index.html`, which is no longer necessary with the current esbuild configuration.